requirements.txt
## -*- coding: utf-8 -*-
# 
# Copyright (c) 2021, LBNL, Alex Weber-Bargioni Group
#
# This file is distributed under the terms
# of the BSD 3-Clause license.
numpy
matplotlib
scipy
gpcam == 6.0.0
pytorch